move code to zookeeper util
