+++ Makefile
@@ -14,6 +14,7 @@ VERSION=$(shell git describe --always --tags --dirty | sed "s/\(.*\)-g`git rev-p
 GIT_SHA=$(shell git rev-parse --short HEAD)
 GOOS=linux
 GOARCH=amd64
+DEPLOY_IMAGE=pravega/pravega-operator:latest
 TEST_IMAGE=tristan1900/test:v0.0.1
 
 .PHONY: all dep build check clean test
@@ -40,7 +41,7 @@ test-unit:
 	go test $$(go list ./... | grep -v /vendor/ | grep -v /test/e2e )
 
 test-e2e:
-	sed "s@REPLACE_IMAGE@$(TEST_IMAGE)@g" -i deploy/operator.yaml
+	sed "s@DEPLOY_IMAGE@$(TEST_IMAGE)@g" -i deploy/operator.yaml
 	operator-sdk build $(TEST_IMAGE) --enable-tests
 	docker login -u "$(DOCKER_TEST_USER)" -p "$(DOCKER_TEST_PASS)"
 	docker push $(TEST_IMAGE)
